+++ junit/kubernetes-server-mock/src/main/java/io/fabric8/kubernetes/client/server/mock/KubernetesCrudDispatcher.java
@@ -41,6 +41,7 @@ import java.net.URISyntaxException;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -83,7 +84,7 @@ public class KubernetesCrudDispatcher extends CrudDispatcher implements Kubernet
     postHandler = new PostHandler(this.kubernetesAttributesExtractor, this);
     putHandler = new PutHandler(this);
     patchHandler = new PatchHandler(this);
-    crdContexts.stream().forEach(this::expectCustomResource);
+    crdContexts.forEach(this::expectCustomResource);
   }
 
   MockResponse process(RecordedRequest request, KubernetesCrudDispatcherHandler handler) {
@@ -207,7 +208,8 @@ public class KubernetesCrudDispatcher extends CrudDispatcher implements Kubernet
     }
     if (!resource.isMarkedForDeletion()) {
       // Mark the resource as deleted, but don't remove it yet (wait for finalizer-removal).
-      resource.getMetadata().setDeletionTimestamp(ZonedDateTime.now(ZoneOffset.UTC).format(DateTimeFormatter.ISO_INSTANT));
+      resource.getMetadata().setDeletionTimestamp(
+          ZonedDateTime.now(ZoneOffset.UTC).truncatedTo(ChronoUnit.SECONDS).format(DateTimeFormatter.ISO_OFFSET_DATE_TIME));
       resource.getMetadata().setResourceVersion(String.valueOf(requestResourceVersion()));
       String updatedResource = Serialization.asJson(resource);
       processEvent(path, pathAttributes, oldAttributes, resource, updatedResource);